requirements.txt
opencv-python
numpy
opencv-python
torch
torchvision
moviepy==1.0.3
Pillow
scipy
matplotlib
